+++ Common/src/main/java/com/leobeliik/extremesoundmuffler/Constants.java
@@ -1,0 +1,6 @@
+package com.leobeliik.extremesoundmuffler;
+
+public class Constants {
+    public static final String MOD_ID = "extremesoundmuffler";
+    public static final String MOD_NAME = "Extreme Sound Muffler";
+}
