+++ Common/src/main/java/com/leobeliik/extremesoundmuffler/MixinConnector.java
@@ -1,0 +1,11 @@
+package com.leobeliik.extremesoundmuffler;
+
+import org.spongepowered.asm.mixin.Mixins;
+import org.spongepowered.asm.mixin.connect.IMixinConnector;
+
+public class MixinConnector implements IMixinConnector {
+    @Override
+    public void connect() {
+        Mixins.addConfiguration("extremesoundmuffler.mixins.json");
+    }
+}
