+++ Common/src/main/java/com/leobeliik/extremesoundmuffler/interfaces/IColorsGui.java
