+++ Common/src/main/java/com/leobeliik/extremesoundmuffler/mixins/InventoryScreenMixin.java
@@ -1,0 +1,26 @@
+package com.leobeliik.extremesoundmuffler.mixins;
+
+import com.leobeliik.extremesoundmuffler.gui.buttons.InvButton;
+import net.minecraft.client.gui.screens.inventory.EffectRenderingInventoryScreen;
+import net.minecraft.client.gui.screens.inventory.InventoryScreen;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.entity.player.Inventory;
+import net.minecraft.world.inventory.InventoryMenu;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(InventoryScreen.class)
+public abstract class InventoryScreenMixin extends EffectRenderingInventoryScreen<InventoryMenu> {
+
+    public InventoryScreenMixin(InventoryMenu inventoryMenu, Inventory inventory, Component component) {
+        super(inventoryMenu, inventory, component);
+    }
+
+    @Inject(method = "init", at = @At("TAIL"))
+    private void InventoryScreenRender(CallbackInfo CI) {
+        System.out.println("AAAAAAAAAAAAAA");
+        this.addRenderableWidget(new InvButton(this, 150, 100/*Config.getInvButtonHorizontal(), Config.getInvButtonVertical()*/));
+    }
+}
