+++ Common/src/main/resources/extremesoundmuffler.mixins.json
@@ -1,0 +1,16 @@
+{
+  "required": true,
+  "minVersion": "0.8",
+  "package": "com.leobeliik.extremesoundmuffler.mixins",
+  "compatibilityLevel": "JAVA_16",
+  "client": [
+    "ClientPlayerMixin",
+    "InventoryScreenMixin",
+    "SoundMixin"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  },
+  "refmap": "extremesoundmuffler.refmap.json"
+}
+
