+++ Fabric/src/main/java/com/leobeliik/extremesoundmuffler/SoundMuffler.java
@@ -1,0 +1,18 @@
+package com.leobeliik.extremesoundmuffler;
+
+import net.fabricmc.api.ModInitializer;
+import net.fabricmc.fabric.api.client.item.v1.ItemTooltipCallback;
+
+public class SoundMuffler implements ModInitializer {
+    
+    @Override
+    public void onInitialize() {
+        
+        // This method is invoked by the Fabric mod loader when it is ready
+        // to load your mod. You can access Fabric and Common code in this
+        // project.
+
+        // Use Fabric to bootstrap the Common mod.
+        CommonClass.init();
+    }
+}
