+++ Forge/src/main/resources/META-INF/MANIFEST.MF
