+++ Forge/src/main/resources/META-INF/mods.toml
@@ -1,0 +1,29 @@
+modLoader="javafml"
+loaderVersion="[34,)"
+license= "GNU LESSER GENERAL PUBLIC LICENSE"
+licenseURL= "https://github.com/LeoBeliik/ExtremeSoundMuffler/blob/master/LICENSE.txt"
+issueTrackerURL="https://github.com/LeoBeliik/ExtremeSoundMuffler/issues"
+[[mods]]
+modId="extremesoundmuffler"
+version="${file.jarVersion}"
+displayName="Extreme Sound Muffler"
+displayURL="https://www.curseforge.com/minecraft/mc-mods/extreme-sound-muffler"
+authors="LeoBeliik"
+logoFile= "esm_logo.png"
+description='''
+This mod allows you to muffle (almost) any sound selectively,
+allowing you to choose the volume of the sound you want between mute and 90%.
+This mod also works Clientside only.
+'''
+[[dependencies.extremeSoundMuffler]]
+    modId="forge"
+    mandatory=true
+    versionRange="[39,)"
+    ordering="NONE"
+    side="BOTH"
+[[dependencies.extremeSoundMuffler]]
+    modId="minecraft"
+    mandatory=true
+    versionRange="[1.18.1,)"
+    ordering="NONE"
+    side="BOTH"
