+++ gradle.properties
@@ -1,7 +1,27 @@
-org.gradle.jvmargs=-Xmx2G
-org.gradle.daemon=false
-
-mcversion = 1.18.1
-forgeversion = 39.0.9
+# Project
 modversion = 19
-mapversion = 1.18
+group=com.leobeliik.extremesoundmuffler
+
+# Common
+minecraft_version=1.18.1
+common_runs_enabled=false
+common_client_run_name=Common Client
+common_server_run_name=Common Server
+
+
+# Forge
+forge_version=39.0.9
+//forge_ats_enabled=true
+
+# Fabric
+fabric_version=0.44.0+1.18
+fabric_loader_version=0.12.12
+
+# Mod options
+mod_name=ExtremeSoundMuffler
+mod_author=LeoBeliik
+mod_id=extremesoundmuffler
+
+# Gradle
+org.gradle.jvmargs=-Xmx3G
+org.gradle.daemon=false
