--- src/main/java/com/leobeliik/extremesoundmuffler/MixinConnector.java
@@ -1,11 +1,0 @@
-package com.leobeliik.extremesoundmuffler;
-
-import org.spongepowered.asm.mixin.Mixins;
-import org.spongepowered.asm.mixin.connect.IMixinConnector;
-
-public class MixinConnector implements IMixinConnector {
-    @Override
-    public void connect() {
-        Mixins.addConfiguration("assets/" + SoundMuffler.MODID + "/extremesoundmuffler.mixins.json");
-    }
-}
