--- src/main/resources/assets/extremesoundmuffler/extremesoundmuffler.mixins.json
@@ -1,15 +1,0 @@
-{
-  "required": true,
-  "package": "com.leobeliik.extremesoundmuffler.mixins",
-  "compatibilityLevel": "JAVA_8",
-  "refmap": "extremesoundmuffler.refmap.json",
-  "minVersion": "0.8",
-  "client": [
-    "SoundMixin",
-    "InventoryScreenMixin",
-    "ClientPlayerMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
