--- src/main/resources/pack.mcmeta
@@ -1,6 +1,0 @@
-{
-    "pack": {
-        "description": "Extreme Sound Muffler resources",
-        "pack_format": 6
-    }
-}
